Renamed gsy_e_core/redis package to gsy_e_core/redis_connections, since Redis package requires the redis name to not be used in packages or modules.
