+++ src/gsy_e/gsy_e_core/redis_connections/aggregator_connection.py
